requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
langchain
langchain-community
langchain-experimental
langchain-huggingface
huggingface_hub
pandas
openpyxl
tabulate
numpy
scikit-learn
matplotlib
seaborn
sentence-transformers